major update for simple decay function
We decided to replace the result made by FPA exponential base 1000 with FPA simple decay; simple decay is as simple as making the expression data for all reactions beyond the distance boundary but keep the un-decayed penalty for reactions within the boundary. In this case, distance boundary is a genuine boundary. All figures related to yeast optimal boundary FPA was updated although the results stayed nearly identical.